refactor: rename repository to highlight it's specifically for the archiver
